video data upload, search, gui change
video data can now be sent and added to the database when a video has been sent successfuly. 
the search service is up and running but not tested.
the gui now has the ability to set the host ip and port number for ease of use. 
removed some unused files